Strike data type fix. int to float

+++ business/entities/core/instruments.go
@@ -13,7 +13,7 @@ type Instrument struct {
 	Token          int64   `json:"token"`
 	InstrumentType string  `json:"instrument_type"`
 	Expiry         string  `json:"expiry"`
-	Strike         int64   `json:"strike"`
+	Strike         float64 `json:"strike"`
 	Price          float64 `json:"price"`
 }
